tests/wpt/web-platform-tests/workers/semantics/run-a-worker/002.html

<!--
var passed = this === self;
onconnect = function(e) {
  e.ports[0].postMessage(passed);
}
/*
-->
<!doctype html>
<title>worker global scope, shared worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new SharedWorker('#');
  worker.port.onmessage = this.step_func(function(e) {
    assert_true(e.data);
    this.done();
  });
});
</script>
<!--
*/
//-->
